Enable passing of build tags
coverall is currently unable to find files that are built conditionally
using build tags. The problem stems from use of the `build.Import`
function which in turn relies on the `build.Default` context. There's
currently no way to configure the `BuildFlags` in this context.

This commit adds a call to `flag.Var()` to ensure that the default
context is configured according to a `-tags` option the same way that
many other go tools do this. See example at
https://godoc.org/golang.org/x/tools/go/buildutil#TagsFlag

Closes #46