Allow specifying the token file instead of the token
It is very helpful to point to a file with a secret token in many cases,
in contrast to an environment variable which somehow needs to be
populated (leads often to exposed secretes in build logs).

Allow passing a token file via the env variable `COVERALLS_TOKEN_FILE` or
via the parameter `-repotokenfile`.

Signed-off-by: Roman Mohr <rmohr@redhat.com>