main: add go-version compatibility.
When a project uses gomodules which are versioned it can happen
that the source file  of the module is not directly under the
module name but instead the versioning happened via a branch so
that the file will be found without the particular version string
as a subdirectory. So this commit will retry to find the particular
file in the main directory instead of the version subdirectory.